+++ src/main/java/insane96mcp/insanelib/InsaneLib.java
@@ -1,11 +1,16 @@
 package insane96mcp.insanelib;
 
+import insane96mcp.insanelib.data.JsonFeatureDataReloadListener;
 import insane96mcp.insanelib.network.NetworkHandler;
 import insane96mcp.insanelib.setup.ClientSetup;
 import insane96mcp.insanelib.setup.Config;
 import insane96mcp.insanelib.setup.ILEntities;
 import insane96mcp.insanelib.setup.ILGlobalLootModifiers;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.AddReloadListenerEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -24,6 +29,7 @@ public class InsaneLib
         ModLoadingContext.get().registerConfig(net.minecraftforge.fml.config.ModConfig.Type.COMMON, Config.COMMON_SPEC);
         FMLJavaModLoadingContext.get().getModEventBus().addListener(ClientSetup::init);
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::preInit);
+        MinecraftForge.EVENT_BUS.register(this);
         final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
         ILEntities.ENTITIES.register(modEventBus);
         ILGlobalLootModifiers.REGISTRY.register(modEventBus);
@@ -31,5 +37,11 @@ public class InsaneLib
 
     public void preInit(FMLCommonSetupEvent event) {
         NetworkHandler.init();
+    }
+
+    @SubscribeEvent(priority = EventPriority.LOW)
+    public void onAddReloadListener(AddReloadListenerEvent event) {
+        JsonFeatureDataReloadListener.reloadContext = event.getConditionContext();
+        event.addListener(JsonFeatureDataReloadListener.INSTANCE);
     }
 }